+++ app/src/main/res/values/dimens.xml
@@ -2,5 +2,5 @@
 <resources>
     <dimen name="dp_8">8dp</dimen>
     <dimen name="dp_4">4dp</dimen>
-    <dimen name="dp_145">145dp</dimen>
+    <dimen name="dp_210">210dp</dimen>
 </resources>
